% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAMD_.R
\name{FAMD_}
\alias{FAMD_}
\title{Factor Analysis of Mixed Data}
\description{
Class of a FAMD object that has fit, fit_transform and transform methods to embed the qualitative and quantitative features into a new quantitative coordinates system
}
\examples{
famd <- FAMD_$new()

X_train_famd <- famd$fit_transform(X_train)

famd$fit(X_train)
X_test_famd <- famd$transform(X_test)
}
% \section{Public fields}{
% \if{html}{\out{<div class="r6-fields">}}
% \describe{
% \item{\code{model}}{FAMD object}

% \item{\code{n_components}}{Number of components to keep from the new coordinates system}
% }
% \if{html}{\out{</div>}}
% }

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FAMD_-new}{\code{FAMD_$new()}}
\item \href{#method-FAMD_-fit}{\code{FAMD_$fit()}}
\item \href{#method-FAMD_-transform}{\code{FAMD_$transform()}}
\item \href{#method-FAMD_-fit_transform}{\code{FAMD_$fit_transform()}}
\item \href{#method-FAMD_-clone}{\code{FAMD_$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-new"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-new}{}}}
\subsection{Method \code{new()}}{
This method initializes a Factor Analys of Mixed Data model by setting up the necessary internal parameters.
It is automatically called when a new object of this class is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$new(n_components = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$fit(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-transform"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$transform(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$fit_transform(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
