% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scaler.R
\name{MinMax.Scaler}
\alias{MinMax.Scaler}
\title{Min-Max Scaler}
\description{
Class of a Min-Max scaler object that has fit, fit_transform and transform methods to min-max scale quantitative data.
}
\examples{
scaler <- MinMax.Scaler$new()

X_train_scaled <- scaler$fit_transform(X_train)

scaler$fit(X_train)
X_test_scaled <- scaler$transform(X_test)

## ------------------------------------------------
## Method `MinMax.Scaler$new`
## ------------------------------------------------

# Creates a new instance of the class
scaler <- MinMax.Scaler$new()

## ------------------------------------------------
## Method `MinMax.Scaler$fit`
## ------------------------------------------------

# Fits the object to a given dataset
scaler$fit(X_train)

## ------------------------------------------------
## Method `MinMax.Scaler$transform`
## ------------------------------------------------

# Performs a min-max scaling on a given dataset
X_test_scaled <- scaler$transform(X_test)

## ------------------------------------------------
## Method `MinMax.Scaler$fit_transform`
## ------------------------------------------------

# Fitting and min-max scaling on a given dataset
X_train_scaled <- scaler$fit_transform(X_train)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{min}}{Vector of feature minimums.}

\item{\code{max}}{Vector of feature maximums.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MinMax.Scaler-new}{\code{MinMax.Scaler$new()}}
\item \href{#method-MinMax.Scaler-fit}{\code{MinMax.Scaler$fit()}}
\item \href{#method-MinMax.Scaler-transform}{\code{MinMax.Scaler$transform()}}
\item \href{#method-MinMax.Scaler-fit_transform}{\code{MinMax.Scaler$fit_transform()}}
\item \href{#method-MinMax.Scaler-clone}{\code{MinMax.Scaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MinMax.Scaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-MinMax.Scaler-new}{}}}
\subsection{Method \code{new()}}{
This method creates a new min-max scaler object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinMax.Scaler$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new min-max scaler object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Creates a new instance of the class
scaler <- MinMax.Scaler$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MinMax.Scaler-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MinMax.Scaler-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits a min-max scaler on a given dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinMax.Scaler$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data on which the scaler computes the features' minimum and maximum.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fits the object to a given dataset
scaler$fit(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MinMax.Scaler-transform"></a>}}
\if{latex}{\out{\hypertarget{method-MinMax.Scaler-transform}{}}}
\subsection{Method \code{transform()}}{
This method applies on a given dataset the transformation rules the scaler learned from the dataset it was fit to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinMax.Scaler$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data that will get min-max scaling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data after min-max scaling.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Performs a min-max scaling on a given dataset
X_test_scaled <- scaler$transform(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MinMax.Scaler-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-MinMax.Scaler-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This method fits a min-max scaler on a given dataset then applies min-max scaling on it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinMax.Scaler$fit_transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data that will get min-max scaling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data after min-max scaling.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fitting and min-max scaling on a given dataset
X_train_scaled <- scaler$fit_transform(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MinMax.Scaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MinMax.Scaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MinMax.Scaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
