% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MultinomialLogisticRegression}
\alias{MultinomialLogisticRegression}
\title{Multinomial Logistic Regression}
\description{
This class is used to perform one-hot encoding on categorical variables in a dataset.
}
\examples{

## ------------------------------------------------
## Method `MultinomialLogisticRegression$new`
## ------------------------------------------------

# Create a new instance of the class
 model <- MultinomialLogisticRegression$new(learning_rate = 0.001,
                                             iterations = 100,
                                              batch_size =4,
                                              regularization = "L1",
                                              regularization_strength = 0.01,
                                              optimizer = "adam",
                                              beta1 = 0.9,
                                              beta2 = 0.999,
                                              epsilon = 1e-8,
                                              lr_decay = TRUE,
                                              decay_rate = 0.01,
                                              warmup_epochs = 10,
                                              initial_learning_rate = 0.0001,
                                              initialization = "zeros")

## ------------------------------------------------
## Method `MultinomialLogisticRegression$fit`
## ------------------------------------------------

model$fit(as.matrix(X_train), y_train)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$predict`
## ------------------------------------------------

predictions <- model$predict(X_test)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$predict_proba`
## ------------------------------------------------

predictions <- model$predict_proba(X_test)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$softmax`
## ------------------------------------------------

probs <- softmax(scores)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$compute_loss`
## ------------------------------------------------

loss <- compute_loss(probs, y_one_hot)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$get_loss_history`
## ------------------------------------------------

loss_history <- model$get_loss_history()

## ------------------------------------------------
## Method `MultinomialLogisticRegression$plot_loss_history`
## ------------------------------------------------

model$plot_loss_history()

## ------------------------------------------------
## Method `MultinomialLogisticRegression$plot_learning_rate_history`
## ------------------------------------------------

model$plot_learning_rate_history()

## ------------------------------------------------
## Method `MultinomialLogisticRegression$get_variable_importance`
## ------------------------------------------------

model$get_variable_importance()

## ------------------------------------------------
## Method `MultinomialLogisticRegression$var_select`
## ------------------------------------------------

model$var_select(5)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$accuracy`
## ------------------------------------------------

model$accuracy(y_test, y_pred)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$confusion_matrix`
## ------------------------------------------------

model$confusion_matrix(y_test, y_pred)

## ------------------------------------------------
## Method `MultinomialLogisticRegression$summary`
## ------------------------------------------------

model$summary()

## ------------------------------------------------
## Method `MultinomialLogisticRegression$print`
## ------------------------------------------------

model$print()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{weights}}{Matrix of weights used by the model for prediction.}

\item{\code{classes}}{Vector of unique class labels derived from the training data.}

\item{\code{loss_history}}{Numeric vector storing the loss values over iterations during training.}

\item{\code{learning_rate_history}}{Numeric vector storing the learning rate values over epochs.}

\item{\code{learning_rate}}{Numeric value indicating the learning rate used for optimization.}

\item{\code{optimizer}}{String specifying the optimization method (e.g., "adam", "sgd").}

\item{\code{regularization}}{Boolean value telling whether to use the regularization term.}

\item{\code{regularization_strength}}{Numeric value for the strength of the regularization term.}

\item{\code{lr_decay}}{Logical value indicating whether learning rate decay is applied.}

\item{\code{decay_rate}}{Numeric value for the learning rate decay factor.}

\item{\code{loss_function}}{String specifying the loss function to use ("cross_entropy" or "squared_error").}

\item{\code{epsilon}}{Numeric value to prevent division by zero in optimization calculations.}

\item{\code{batch_size}}{Integer specifying the number of samples per mini-batch for training.}

\item{\code{beta1}}{Numeric value for the exponential decay rate for the first moment in the Adam optimizer.}

\item{\code{beta2}}{Numeric value for the exponential decay rate for the second moment in the Adam optimizer.}

\item{\code{m}}{Matrix used for storing the first moment estimates in the Adam optimizer.}

\item{\code{v}}{Matrix used for storing the second moment estimates in the Adam optimizer.}

\item{\code{t}}{Integer representing the current iteration for the Adam optimizer.}

\item{\code{iterations}}{Integer specifying the total number of iterations for training.}

\item{\code{warmup_epochs}}{Integer specifying the number of warmup epochs where the learning rate is gradually increased.}

\item{\code{initial_learning_rate}}{Numeric value for the initial learning rate during the warmup phase.}

\item{\code{initialization}}{String specifying the method used for initializing weights ("xavier" or "zeros").}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultinomialLogisticRegression-new}{\code{MultinomialLogisticRegression$new()}}
\item \href{#method-MultinomialLogisticRegression-fit}{\code{MultinomialLogisticRegression$fit()}}
\item \href{#method-MultinomialLogisticRegression-predict}{\code{MultinomialLogisticRegression$predict()}}
\item \href{#method-MultinomialLogisticRegression-predict_proba}{\code{MultinomialLogisticRegression$predict_proba()}}
\item \href{#method-MultinomialLogisticRegression-softmax}{\code{MultinomialLogisticRegression$softmax()}}
\item \href{#method-MultinomialLogisticRegression-compute_loss}{\code{MultinomialLogisticRegression$compute_loss()}}
\item \href{#method-MultinomialLogisticRegression-get_loss_history}{\code{MultinomialLogisticRegression$get_loss_history()}}
\item \href{#method-MultinomialLogisticRegression-plot_loss_history}{\code{MultinomialLogisticRegression$plot_loss_history()}}
\item \href{#method-MultinomialLogisticRegression-plot_learning_rate_history}{\code{MultinomialLogisticRegression$plot_learning_rate_history()}}
\item \href{#method-MultinomialLogisticRegression-get_variable_importance}{\code{MultinomialLogisticRegression$get_variable_importance()}}
\item \href{#method-MultinomialLogisticRegression-var_select}{\code{MultinomialLogisticRegression$var_select()}}
\item \href{#method-MultinomialLogisticRegression-accuracy}{\code{MultinomialLogisticRegression$accuracy()}}
\item \href{#method-MultinomialLogisticRegression-confusion_matrix}{\code{MultinomialLogisticRegression$confusion_matrix()}}
\item \href{#method-MultinomialLogisticRegression-summary}{\code{MultinomialLogisticRegression$summary()}}
\item \href{#method-MultinomialLogisticRegression-print}{\code{MultinomialLogisticRegression$print()}}
\item \href{#method-MultinomialLogisticRegression-clone}{\code{MultinomialLogisticRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-new}{}}}
\subsection{Method \code{new()}}{
This method initializes a model with all defaults paramaters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$new(
  learning_rate = 0.01,
  iterations = 100,
  optimizer = "adam",
  regularization = NULL,
  regularization_strength = 0.01,
  lr_decay = FALSE,
  decay_rate = 0.001,
  loss_function = "cross_entropy",
  batch_size = 32,
  beta1 = 0.9,
  beta2 = 0.999,
  epsilon = 1e-08,
  warmup_epochs = 0,
  initial_learning_rate = 0.001,
  initialization = "xavier"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{Numeric. The learning rate used for optimization.}

\item{\code{iterations}}{Integer. The total number of iterations for training.}

\item{\code{optimizer}}{String. The optimization algorithm to use (e.g., "adam", "sgd").}

\item{\code{regularization}}{String. The type of regularization to apply ("L1", "L2", or NULL).}

\item{\code{regularization_strength}}{Numeric. The strength of the regularization term.}

\item{\code{lr_decay}}{Logical. Whether to apply learning rate decay.}

\item{\code{decay_rate}}{Numeric. The decay rate factor for learning rate decay.}

\item{\code{loss_function}}{String. The loss function to use ("cross_entropy" or "squared_error").}

\item{\code{batch_size}}{Integer. The number of samples per mini-batch for training.}

\item{\code{beta1}}{Numeric. The exponential decay rate for the first moment in the optimizer.}

\item{\code{beta2}}{Numeric. The exponential decay rate for the second moment in the optimizer.}

\item{\code{epsilon}}{Numeric. A small value to prevent division by zero in optimization calculations.}

\item{\code{warmup_epochs}}{Integer. The number of warmup epochs during which the learning rate is gradually increased.}

\item{\code{initial_learning_rate}}{Numeric. The initial learning rate for the warmup phase.}

\item{\code{initialization}}{String. The method used for initializing weights ("xavier" or "zeros").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Initializes an instance of the \code{MultinomialLogisticRegression} class with the specified parameters.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Create a new instance of the class
 model <- MultinomialLogisticRegression$new(learning_rate = 0.001,
                                             iterations = 100,
                                              batch_size =4,
                                              regularization = "L1",
                                              regularization_strength = 0.01,
                                              optimizer = "adam",
                                              beta1 = 0.9,
                                              beta2 = 0.999,
                                              epsilon = 1e-8,
                                              lr_decay = TRUE,
                                              decay_rate = 0.01,
                                              warmup_epochs = 10,
                                              initial_learning_rate = 0.0001,
                                              initialization = "zeros")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits the model to data given.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$fit(
  X,
  y,
  epochs = self$iterations,
  progress_callback = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix. Explanatory variables}

\item{\code{y}}{Factor. Target Variable}

\item{\code{epochs}}{Integer. Iteration for convergence}

\item{\code{progress_callback}}{Function. To view the progress of model training on data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$fit(as.matrix(X_train), y_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-predict}{}}}
\subsection{Method \code{predict()}}{
This function calculates the class predicted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix. Explanatory variables from the test dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector. The predicted class for each value of the test dataset.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{predictions <- model$predict(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
This function calculates the probabilites of belonging to each class of the target variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$predict_proba(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix. Explanatory variables from the test dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix. The probability of belonging to each class.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{predictions <- model$predict_proba(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-softmax"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-softmax}{}}}
\subsection{Method \code{softmax()}}{
This function computes the softmax probabilities for a given set of scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$softmax(scores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scores}}{Matrix. Matrix product between weights and the matrix with the explanatory variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix. Probabilities where each row  sums to 1.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{probs <- softmax(scores)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-compute_loss"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-compute_loss}{}}}
\subsection{Method \code{compute_loss()}}{
This function computes the loss value for a given set of probabilities and one-hot encoded target values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$compute_loss(probs, y_one_hot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{Matrix. Probabilities of belonging to each class.}

\item{\code{y_one_hot}}{Matrix. One-hot encoded target values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Float. The loss value calculated based on the probabilities and target values.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{loss <- compute_loss(probs, y_one_hot)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-get_loss_history"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-get_loss_history}{}}}
\subsection{Method \code{get_loss_history()}}{
This function returns the loss history of the model during training.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$get_loss_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector. Loss values over iterations during training.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{loss_history <- model$get_loss_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-plot_loss_history"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-plot_loss_history}{}}}
\subsection{Method \code{plot_loss_history()}}{
To plot the loss curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$plot_loss_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
lineplot. Loss variations as a function of iteration.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$plot_loss_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-plot_learning_rate_history"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-plot_learning_rate_history}{}}}
\subsection{Method \code{plot_learning_rate_history()}}{
To plot the learning rate by epoch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$plot_learning_rate_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
lineplot. Learning rate by epoch.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$plot_learning_rate_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-get_variable_importance"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-get_variable_importance}{}}}
\subsection{Method \code{get_variable_importance()}}{
To have the importance of each variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$get_variable_importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DataFrame with the importance calculated for each variable.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$get_variable_importance()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-var_select"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-var_select}{}}}
\subsection{Method \code{var_select()}}{
To have the N most important variable in the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$var_select(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Integer. Number of the best variable to select}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector. Names of the n best variable based on their importance.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$var_select(5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-accuracy"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-accuracy}{}}}
\subsection{Method \code{accuracy()}}{
Calculate the accuracy of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$accuracy(y_test, y_pred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_test}}{Factor. The target variable with the real class.}

\item{\code{y_pred}}{Factor. The class predicted from the model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Float. The accuracy of the model.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$accuracy(y_test, y_pred)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-confusion_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-confusion_matrix}{}}}
\subsection{Method \code{confusion_matrix()}}{
Calculate the confusion matrix between the prediction of the model and the real class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$confusion_matrix(y_test, predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_test}}{Factor. The target variable with the real class.}

\item{\code{predictions}}{Factor. The class predicted from the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix. Display the confusion matrix calculated.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$confusion_matrix(y_test, y_pred)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-summary"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-summary}{}}}
\subsection{Method \code{summary()}}{
Display all informations parameters about the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Text. print informations
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$summary()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-print}{}}}
\subsection{Method \code{print()}}{
display a small summary of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Text. print informations
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultinomialLogisticRegression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultinomialLogisticRegression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultinomialLogisticRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
