% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoder.R
\name{One.Hot.Encoder}
\alias{One.Hot.Encoder}
\title{One-Hot Encoder}
\description{
Class of a One-Hot Encoder object that has fit, fit_transform and transform methods to use over the qualitative features you want to encode.
}
\examples{
ohe <- One.Hot.Encoder$new().

X_train_ohe <- ohe$fit_transform(X_train).

ohe$fit(X_train).
X_test_ohe <- ohe$transform(X_test)
}
% \section{Public fields}{
% \if{html}{\out{<div class="r6-fields">}}
% \describe{
% \item{\code{categories}}{(\code{list()}) \cr
% Categories of qualitative features.}
% }
% \if{html}{\out{</div>}}
% }
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-One.Hot.Encoder-new}{\code{One.Hot.Encoder$new()}}
\item \href{#method-One.Hot.Encoder-fit}{\code{One.Hot.Encoder$fit()}}
\item \href{#method-One.Hot.Encoder-transform}{\code{One.Hot.Encoder$transform()}}
\item \href{#method-One.Hot.Encoder-fit_transform}{\code{One.Hot.Encoder$fit_transform()}}
\item \href{#method-One.Hot.Encoder-clone}{\code{One.Hot.Encoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One.Hot.Encoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-One.Hot.Encoder-new}{}}}
\subsection{Method \code{new()}}{
This method initializes a one-hot encoder object by setting up the necessary internal parameters.
It is automatically called when a new object of this class is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One.Hot.Encoder$new()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Creates a new instance of the class
ohe <- One.Hot.Encoder$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One.Hot.Encoder-fit"></a>}}
\if{latex}{\out{\hypertarget{method-One.Hot.Encoder-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits a One-Hot Encoder (OHE) on a given dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One.Hot.Encoder$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_quali}}{A dataframe containing the qualitative data from which the object learns to perform future one-hot encoding.}
}
\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fits the object to a given dataset
ohe$fit(X_train)
	
}
\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One.Hot.Encoder-transform"></a>}}
\if{latex}{\out{\hypertarget{method-One.Hot.Encoder-transform}{}}}
\subsection{Method \code{transform()}}{
This method transforms a given dataset based on that which the OHE was fit to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One.Hot.Encoder$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_quali}}{A dataframe containing the qualitative data that will undergo one-hot encoding.}
}
\if{html}{\out{</div>}}
}

\subsection{Returns}{
A (sparse) dataframe containing the one-hot encoded qualitative features.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# One-hot encoding a given dataset
X_test_ohe <- ohe$transform(X_test)
	
}
\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One.Hot.Encoder-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-One.Hot.Encoder-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This method fits a One-Hot Encoder (OHE) on a given dataset then transforms it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One.Hot.Encoder$fit_transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_quali}}{A dataframe containing the qualitative data to which the object will be fit and perform one-hot encoding.}
}
\if{html}{\out{</div>}}
}

\subsection{Returns}{
A (sparse) dataframe containing the one-hot encoded qualitative features.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fitting and one-hot encoding a given dataset
X_train_ohe <- ohe$fit_transform(X_train)
	
}
\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One.Hot.Encoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-One.Hot.Encoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One.Hot.Encoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
