% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_scaler.R
\name{Robust.Scaler}
\alias{Robust.Scaler}
\title{Robust (Median) Scaler}
\description{
Class of a robust (median) scaler object that has fit, fit_transform and transform methods to scale quantitative data.
}
\examples{
scaler <- Robust.Scaler$new()

X_train_scaled <- scaler$fit_transform(X_train)

scaler$fit(X_train)
X_test_scaled <- scaler$transform(X_test)

## ------------------------------------------------
## Method `Robust.Scaler$new`
## ------------------------------------------------

# Creates a new instance of the class
scaler <- Robust.Scaler$new()

## ------------------------------------------------
## Method `Robust.Scaler$fit`
## ------------------------------------------------

# Fits the object to a given dataset
scaler$fit(X_train)

## ------------------------------------------------
## Method `Robust.Scaler$transform`
## ------------------------------------------------

# Performs a robust scaling on a given dataset
X_test_scaled <- scaler$transform(X_test)

## ------------------------------------------------
## Method `Robust.Scaler$fit_transform`
## ------------------------------------------------

# Fitting and robust scaling on a given dataset
X_train_scaled <- scaler$fit_transform(X_train)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{median}}{Vector of feature medians.}

\item{\code{iqr}}{Vector of feature inter-quartile ranges.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Robust.Scaler-new}{\code{Robust.Scaler$new()}}
\item \href{#method-Robust.Scaler-fit}{\code{Robust.Scaler$fit()}}
\item \href{#method-Robust.Scaler-transform}{\code{Robust.Scaler$transform()}}
\item \href{#method-Robust.Scaler-fit_transform}{\code{Robust.Scaler$fit_transform()}}
\item \href{#method-Robust.Scaler-clone}{\code{Robust.Scaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Robust.Scaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-Robust.Scaler-new}{}}}
\subsection{Method \code{new()}}{
This method creates a new robust scaler object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Robust.Scaler$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new standard robust object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Creates a new instance of the class
scaler <- Robust.Scaler$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Robust.Scaler-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Robust.Scaler-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits a robust scaler on a given dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Robust.Scaler$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data on which the scaler computes the features' median and inter-quartile range.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fits the object to a given dataset
scaler$fit(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Robust.Scaler-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Robust.Scaler-transform}{}}}
\subsection{Method \code{transform()}}{
This method applies on a given dataset the transformation rules the scaler learned from the dataset it was fit to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Robust.Scaler$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data that will get robust scaling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data after robust scaling.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Performs a robust scaling on a given dataset
X_test_scaled <- scaler$transform(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Robust.Scaler-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Robust.Scaler-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This method fits a robust scaler on a given dataset then applies robust scaling on it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Robust.Scaler$fit_transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative data that will get robust scaling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data after robust scaling.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fitting and robust scaling on a given dataset
X_train_scaled <- scaler$fit_transform(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Robust.Scaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Robust.Scaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Robust.Scaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
