% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_quanti_imputer.R
\name{Simple.Quanti.Imputer}
\alias{Simple.Quanti.Imputer}
\title{Simple quantitative NA imputer}
\description{
Class of a quantitative NA imputer object that has fit, fit_transform and transform methods to impute quantitative missing values with mean or median method
}
\examples{
imputer <- Simple.Quanti.Imputer$new(imputation_type="mean")
imputer <- Simple.Quanti.Imputer$new(imputation_type="median")

X_train_filled <- imputer$fit_transform(X_train, c("Col_quanti_1","Col_quanti_2","Col_quanti_N"))

imputer$fit(X_train, c("Col_quanti_1","Col_quanti_2","Col_quanti_N"))
X_test_filled <- imputer$transform(X_test)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{imputation_type}}{Method that will be used to impute NA}

\item{\code{imputation_values}}{Values produced by the imputer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simple.Quanti.Imputer-new}{\code{Simple.Quanti.Imputer$new()}}
\item \href{#method-Simple.Quanti.Imputer-fit}{\code{Simple.Quanti.Imputer$fit()}}
\item \href{#method-Simple.Quanti.Imputer-transform}{\code{Simple.Quanti.Imputer$transform()}}
\item \href{#method-Simple.Quanti.Imputer-fit_transform}{\code{Simple.Quanti.Imputer$fit_transform()}}
\item \href{#method-Simple.Quanti.Imputer-clone}{\code{Simple.Quanti.Imputer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quanti.Imputer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quanti.Imputer-new}{}}}
\subsection{Method \code{new()}}{
Create a new simple quantitative imputer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quanti.Imputer$new(imputation_type = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imputation_type}}{Method of NA imputation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new imputer object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quanti.Imputer-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quanti.Imputer-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the simple quantitative imputer to data following the specified method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quanti.Imputer$fit(data, columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}

\item{\code{columns}}{Vector of column names that will get imputation (must be of quantitative type)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quanti.Imputer-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quanti.Imputer-transform}{}}}
\subsection{Method \code{transform()}}{
Transform the passed data object based on the fit data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quanti.Imputer$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe of data after imputation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quanti.Imputer-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quanti.Imputer-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Fit to data object following the specified method then impute NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quanti.Imputer$fit_transform(data, columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}

\item{\code{columns}}{Vector of column names that will get imputation (must be of quantitative type)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe of data after imputation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quanti.Imputer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quanti.Imputer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quanti.Imputer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
