% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{split_train_test}
\alias{split_train_test}
\title{Split train-test}
\usage{
split_train_test(
  data,
  target_var,
  explanatory_vars,
  train_ratio = 0.8,
  random_state = NULL,
  shuffle = TRUE
)
}
\arguments{
\item{data}{Dataframe with all the data without any modfications.}

\item{target_var}{String or List with the target variable}

\item{explanatory_vars}{Matrix with all the explanatory variable}

\item{train_ratio}{Float. Percentage to be passed in the sample train.}

\item{shuffle}{binary. If true, then the samples are constructed with randomness.}

\item{random_stat}{Integer. Controls the random seed for reproducibility.}
}
\value{
List with data separated into trains/tests
}
\description{
To split the dataset with train/test method
}
\examples{
split_data <- split_train_test(data = data, target_var = target_var, explanatory_vars = explanatory_vars, train_ratio = 0.9, random_state = 123,
}
