% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_scaler.R
\name{Standard.Scaler}
\alias{Standard.Scaler}
\title{Standardization Scaler}
\description{
Class of a Standardization scaler object that has fit, fit_transform and transform methods to standardize quantitative data
}
\examples{
scaler <- Standard.Scaler$new()

X_train_scaled <- scaler$fit_transform(X_train)

scaler$fit(X_train)
X_test_scaled <- scaler$transform(X_test)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mean}}{Vector of feature means}

\item{\code{std}}{Vector of feature standard-deviations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Standard.Scaler-new}{\code{Standard.Scaler$new()}}
\item \href{#method-Standard.Scaler-fit}{\code{Standard.Scaler$fit()}}
\item \href{#method-Standard.Scaler-transform}{\code{Standard.Scaler$transform()}}
\item \href{#method-Standard.Scaler-fit_transform}{\code{Standard.Scaler$fit_transform()}}
\item \href{#method-Standard.Scaler-clone}{\code{Standard.Scaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Standard.Scaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-Standard.Scaler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Standard.Scaler$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Standard.Scaler-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Standard.Scaler-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Standard.Scaler$fit(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Standard.Scaler-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Standard.Scaler-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Standard.Scaler$transform(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Standard.Scaler-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Standard.Scaler-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Standard.Scaler$fit_transform(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Standard.Scaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Standard.Scaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Standard.Scaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
