% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAMD_.R
\name{FAMD_}
\alias{FAMD_}
\title{Factor Analysis of Mixed Data (FAMD)}
\description{
Class of a Factorial Analysis of Mixed Data (FAMD) object that has fit, fit_transform and transform methods to embed the qualitative and quantitative features into a new quantitative coordinates system.
}
\examples{
famd <- FAMD_$new(n_components = 2)

X_train_famd <- famd$fit_transform(X_train)

famd$fit(X_train)
X_test_famd <- famd$transform(X_test)

## ------------------------------------------------
## Method `FAMD_$new`
## ------------------------------------------------

# Creates a new instance of the class
famd <- FAMD_$new(n_components = 2)

## ------------------------------------------------
## Method `FAMD_$fit`
## ------------------------------------------------

# Fits the object to a given dataset
famd$fit(X_train)

## ------------------------------------------------
## Method `FAMD_$transform`
## ------------------------------------------------

# Performs a FAMD on a given dataset
X_test_famd <- famd$transform(X_test)

## ------------------------------------------------
## Method `FAMD_$fit_transform`
## ------------------------------------------------

# Fitting and FAMD modelling a given dataset
X_train_famd <- famd$fit_transform(X_train)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The FAMD object.}

\item{\code{n_components}}{Number of components to keep from the FAMD factorial space.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FAMD_-new}{\code{FAMD_$new()}}
\item \href{#method-FAMD_-fit}{\code{FAMD_$fit()}}
\item \href{#method-FAMD_-transform}{\code{FAMD_$transform()}}
\item \href{#method-FAMD_-fit_transform}{\code{FAMD_$fit_transform()}}
\item \href{#method-FAMD_-clone}{\code{FAMD_$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-new"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-new}{}}}
\subsection{Method \code{new()}}{
This method initializes a FAMD model by setting up the necessary internal parameters.
It is automatically called when a new object of this class is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$new(n_components = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_components}}{Number of components to keep from the FAMD factorial space.
When no value is specified, it is automatically set to \code{ncol(data)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new FAMD model object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Creates a new instance of the class
famd <- FAMD_$new(n_components = 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits a FAMD model on a given dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the qualitative data from which the object learns FAMD data representation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fits the object to a given dataset
famd$fit(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-transform"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-transform}{}}}
\subsection{Method \code{transform()}}{
This method applies on a given dataset the transformation rules the FAMD model learned from the dataset it was fit to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative and qualitative data that will undergo FAMD modelisation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the coordinates of the new data representation in the FAMD factorial space.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Performs a FAMD on a given dataset
X_test_famd <- famd$transform(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This method fits a FAMD model on a given dataset then transforms it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$fit_transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the quantitative and qualitative data to which the object will be fit and will perform FAMD.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the coordinates of the new data representation in the FAMD factorial space.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fitting and FAMD modelling a given dataset
X_train_famd <- famd$fit_transform(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FAMD_-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FAMD_-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FAMD_$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
