% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_quali_imputer.R
\name{Simple.Quali.Imputer}
\alias{Simple.Quali.Imputer}
\title{Simple qualitative NA imputer}
\description{
Class of a qualitative NA imputer object that has fit, fit_transform and transform methods to impute qualitative missing values with remove, mode or probabilistic method.
}
\examples{
imputer <- Simple.Quali.Imputer$new(imputation_type="mode")
imputer <- Simple.Quali.Imputer$new(imputation_type="probability")

X_train_filled <- imputer$fit_transform(X_train, c("Col_quali_1","Col_quali_2","Col_quali_N"))

imputer$fit(X_train, c("Col_quali_1","Col_quali_2","Col_quali_N"))
X_test_filled <- imputer$transform(X_test)

## ------------------------------------------------
## Method `Simple.Quali.Imputer$new`
## ------------------------------------------------

# Creates a new instance of the class
imputer <- Simple.Quali.Imputer$new(imputation_type = "probability")

## ------------------------------------------------
## Method `Simple.Quali.Imputer$fit`
## ------------------------------------------------

# Fits the object to a given dataset
imputer$fit(X_train)

## ------------------------------------------------
## Method `Simple.Quali.Imputer$transform`
## ------------------------------------------------

# Performs a qualitative imputation on a given dataset
X_test_filled <- imputer$transform(X_test)

## ------------------------------------------------
## Method `Simple.Quali.Imputer$fit_transform`
## ------------------------------------------------

# Fitting and imputing missing values on a given dataset
X_train_filled <- imputer$fit_transform(X_train)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{imputation_type}}{Method that will be used to impute NA.}

\item{\code{imputation_values}}{Values produced by the imputer.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simple.Quali.Imputer-new}{\code{Simple.Quali.Imputer$new()}}
\item \href{#method-Simple.Quali.Imputer-fit}{\code{Simple.Quali.Imputer$fit()}}
\item \href{#method-Simple.Quali.Imputer-transform}{\code{Simple.Quali.Imputer$transform()}}
\item \href{#method-Simple.Quali.Imputer-fit_transform}{\code{Simple.Quali.Imputer$fit_transform()}}
\item \href{#method-Simple.Quali.Imputer-clone}{\code{Simple.Quali.Imputer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quali.Imputer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quali.Imputer-new}{}}}
\subsection{Method \code{new()}}{
This method creates a new qualitative imputer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quali.Imputer$new(imputation_type = "mode")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imputation_type}}{String defining the method of NA imputation.
When it equals \code{mode}, the object is set to perform an imputation based on the mode, i.e. the most frequent modality of the feature.
When it equals \code{probability}, the object is set to perform an imputation based on the probability distribution based on the modality frequencies, i.e. each NA is imputed by one of the possible feature modalities based on a probability weighted by its frequency of appearance.
When it equals \code{remove}, the object is set to perform observations removal if they have missing values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new qualitative imputer object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Creates a new instance of the class
imputer <- Simple.Quali.Imputer$new(imputation_type = "probability")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quali.Imputer-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quali.Imputer-fit}{}}}
\subsection{Method \code{fit()}}{
This method fits a qualitative imputer to data given the specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quali.Imputer$fit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the qualitative data on which the imputer learns to fill missing values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing. The object is internally updated when using this method.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fits the object to a given dataset
imputer$fit(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quali.Imputer-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quali.Imputer-transform}{}}}
\subsection{Method \code{transform()}}{
This method applies on a given dataset the transformation rules the imputer learned from the dataset it was fit to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quali.Imputer$transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the qualitative data that will get missing values imputation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data with missing values filled.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Performs a qualitative imputation on a given dataset
X_test_filled <- imputer$transform(X_test)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quali.Imputer-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quali.Imputer-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This method fits an imputer on a given dataset then fill the missing values in it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quali.Imputer$fit_transform(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the qualitative data that will get missing values imputation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the data with its missing values filled.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Fitting and imputing missing values on a given dataset
X_train_filled <- imputer$fit_transform(X_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simple.Quali.Imputer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simple.Quali.Imputer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simple.Quali.Imputer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
